# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Xihe API
  version: 1.0.0
  description: Canonical HTTP contract for Xihe public and service APIs.
servers:
  - url: /
paths:
...
  /api/v1/sessions:
    get: { operationId: listSessions, security: [{ bearerAuth: [] }], responses: { '200': { description: Sessions } } }
    post: { operationId: createSession, security: [{ bearerAuth: [] }], responses: { '201': { description: Created }, '401': { $ref: '#/components/responses/Problem' } } }
  /api/v1/sessions/{sessionId}:
    delete: { operationId: deleteSession, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
...
  /internal/v1/config/{layer}/{domain}:
    get:
      operationId: getInternalConfig
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: layer, required: true, schema: { type: string } }, { in: path, name: domain, required: true, schema: { type: string } }]
      responses: { '200': { description: Configuration map }, '401': { $ref: '#/components/responses/Problem' } }
...
  /api/v1/config/{domain}:
    get: { operationId: getConfig, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: domain, required: true, schema: { type: string } }], responses: { '200': { description: Configuration map } } }
  /api/v1/config/admin/{domain}:
    put: { operationId: putAdminConfig, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: domain, required: true, schema: { type: string } }], responses: { '200': { description: Updated } } }
  /api/v1/config/user/{domain}:
    put: { operationId: putUserConfig, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: domain, required: true, schema: { type: string } }], responses: { '200': { description: Updated } } }
  /api/v1/config/import:
    post: { operationId: importConfig, security: [{ bearerAuth: [] }], responses: { '200': { description: Imported } } }
  /api/v1/config/export:
    get: { operationId: exportConfig, security: [{ bearerAuth: [] }], responses: { '200': { description: JSON configuration } } }
  /api/v1/providers:
    get: { operationId: listProviders, security: [{ bearerAuth: [] }], responses: { '200': { description: Providers } } }
    put: { operationId: updateProviders, security: [{ bearerAuth: [] }], responses: { '200': { description: Updated } } }
...
  /api/v1/sessions/{sessionId}/messages:
    get: { operationId: listMessages, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }], responses: { '200': { description: Messages } } }
  /api/v1/sessions/{sessionId}/messages/{messageId}:
    delete: { operationId: deleteMessage, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: sessionId, required: true, schema: { type: string } }, { in: path, name: messageId, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
...
  /api/v1/workspaces/{workspaceId}/mcp-config:
     get: { operationId: getMcpConfig, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: workspaceId, required: true, schema: { type: string } }], responses: { '200': { description: MCP configuration } } }
     put: { operationId: putMcpConfig, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: workspaceId, required: true, schema: { type: string } }], responses: { '200': { description: Updated } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  schemas:
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        requestId: { type: string }